requirements.txt
streamlit
roboticstoolbox-python
pandas
numpy
watchdog
matplotlib
pytest
pynput
tqdm